refactor(plugin-tower): move EncodeRequest to request.go
